aggregator/requirements.txt
flask
flask-cors
kafka-python
python-dotenv